Extract vm creation to allow shared execution context
